+++ student.rb
@@ -13,6 +13,23 @@ class Student < Person
     '¯(ツ)/¯'
   end
 
+  def self.create_student
+    print 'Name: '
+    name = gets.chomp
+    print 'Classroom: '
+    classroom = gets.chomp.to_str
+    return nil unless classroom.length >= 1
+
+    print 'Age (numeric): '
+    age = gets.chomp.to_i
+    return nil unless age.positive?
+
+    print 'Parent permission (y for true or n for false): '
+    parent_permission = gets.chomp
+
+    Student.new(classroom: classroom, age: age, name: name, parent_permission: parent_permission.downcase == 'y')
+  end
+
   def add_to_classroom(classroom)
     @classroom = classroom
     classroom.students.push(self) unless classroom.students.include?(self)
